[DO NOT MERGE] Use all tracked probabilities to estimate a quantile
This commit puts accumulators as a public attribute of accumulator_set, and heights as a public attribute of weighted_extended_p_square to make use of it in demonstrating all heights of size 2 * probabilities + 3 (going from 0 to 1 through each half-way point and probability) can be used to estimate (e.g. in linear interpolation) a new quantile.
This is in contrast of boost::accumulators::quantile(accumulator_set, quantile) which relies on extractor<extended_p_square_tag> to retrieve only the heights associated with provided probabilities. By doing that, more than half the information that is computed by P-SQUARE is thrown away and it can be shown (see # ) that interpolating with this additional information already computed in boost but unused as of now yields better results.
However, a proper solution would be to implement this inside the boost::extractor mechanism, which is out of the scope of this commit as it may impact other features in accumulators.